Add option defer-permissions for OS X
This option is only available on OS X and disables local permission
checks and forwards all requests to the remote server.

The problem is that OS X does not understand the groups that it
receives from the server and disallows access because your user does
not belong to any groups.

Check https://github.com/osxfuse/filesystems/issues/1#issuecomment-3612902
for a more in-depth explanation.